[ENH] budget_constraint: check constraint on budget_commit function
