+++ budget_constraint/models/__init__.py
@@ -1,4 +1,4 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import budget_period
+from . import base_budget_move
 from . import budget_constraint
