--- budget_constraint/views/budget_period_view.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="budget_period_view_form" model="ir.ui.view">
-        <field name="name">budget.period.view.form</field>
-        <field name="model">budget.period</field>
-        <field name="inherit_id" ref="budget_control.budget_period_view_form" />
-        <field name="arch" type="xml">
-            <field name="control_all_analytic_accounts" position="before">
-                <field name="use_budget_constraint" />
-            </field>
-        </field>
-    </record>
-</odoo>
